Finished separating showClass.
